background: Change title match condition
When play-opening a new playlist from e.g. the home page, relying on
the URL changed event leaves a race between customize.js and the
native js starting the music. If we match for the exact title
condition we care about, we'll simply re-trigger customize.js if the
race didn't work in our favor.